+++ .gitignore
@@ -3,5 +3,6 @@ node_modules
 *.log*
 dist
 /runtime
+/web
 .tmp
 .wrangler
