+++ package.json
@@ -17,13 +17,19 @@
       "types": "./runtime/*.d.ts",
       "require": "./runtime/*.cjs",
       "import": "./runtime/*.mjs"
+    },
+    "./web/*": {
+      "types": "./web/*.d.ts",
+      "require": "./web/*.cjs",
+      "import": "./web/*.mjs"
     }
   },
   "main": "./dist/index.cjs",
   "types": "./dist/index.d.ts",
   "files": [
     "dist",
-    "runtime"
+    "runtime",
+    "web"
   ],
   "scripts": {
     "build": "unbuild",
